Add script that loads app states and interactive tutorials into BigQuery
Maria produced the `app-states.json` file from https://github.com/grafana/docs-plugin/blob/product-knowledge/product-knowledge/app-states.mdc.
I've committed it for reproducibility.

I've left the create code there for posterity but script doesn't attempt to create the tables idempotently and that command will fail if re-run.
If we ever need to change that in the future it should be trivial.

The update command replaces the data in the table with the contents of the `app-states.json` and the raw interactive tutorials `index.json` fetched from GitHub.

Anyone can run this script if they have the appropriate BigQuery permissions but you may have to first `gcloud auth login` and you may also have to install `bq` with `gcloud components install bq`.

Signed-off-by: Jack Baldry <jack.baldry@grafana.com>